requirements.txt
streamlit==1.28.1
pandas==2.1.1
openpyxl==3.1.2
pytesseract==0.3.10
Pillow==10.0.1
pdf2image==1.16.3
PyMuPDF==1.23.5
python-docx==0.8.11
docx2txt==0.8
numpy==1.24.3
requests==2.31.0